requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-migrate==4.0.4
email-validator==2.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
flask-cors==4.0.0
